fix : 타입 오류 해결

+++ src/app/(provider)/(root)/concerts/[postId]/page.tsx
@@ -1,10 +1,13 @@
 "use client";
 
 import ConcertDeleteButton from "@/components/ConcertList/ConcertDeleteButton";
+import { Concert, ConcertInDB } from "@/types/Concert";
 import { formatDateString } from "@/utils/formatDateString";
 import { createClient } from "@/utils/supabase/client";
+import { User } from "@supabase/supabase-js";
 import { useQuery } from "@tanstack/react-query";
 import moment from "moment";
+import Image from "next/image";
 import Link from "next/link";
 import { useEffect, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
@@ -14,7 +17,7 @@ type ConcertDetailPageProps = {
 };
 
 const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
-  const [user, setUser] = useState(null);
+  const [user, setUser] = useState<User>();
   const [like, setLike] = useState(0);
   const [heart, setHeart] = useState(false);
 
@@ -22,7 +25,9 @@ const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
   useEffect(() => {
     const fetchData = async () => {
       const { data, error: getUserError } = await supabase.auth.getUser();
-      setUser(data.user);
+      if (data.user) {
+        setUser(data.user);
+      }
     };
     fetchData();
   }, []);
@@ -34,7 +39,9 @@ const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
         .from("concert_likes")
         .select()
         .eq("post_id", postId);
-      setLike(data.length);
+      if (data) {
+        setLike(data.length);
+      }
     };
     fetchLike();
   }, [like]);
@@ -45,18 +52,18 @@ const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
       const supabase = createClient();
       const { data, error } = await supabase
         .from("concert_posts")
-        .select("*, users:author_id(*)")
+        .select("*, users:author_id(nickname, profile_image)")
         .eq("post_id", postId)
         .single();
 
-      return data;
+      return data as any;
     },
   });
 
-  if (isPending) {
+  if (isPending || !concert) {
     return <div>로딩 중입니다...</div>;
   }
-
+  console.log(concert);
   const {
     post_id: id,
     title,
@@ -95,7 +102,7 @@ const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
       .select()
       .eq("post_id", postId)
       .eq("user_id", user.id);
-    if (data.length) {
+    if (data && data.length) {
       await supabase
         .from("concert_likes")
         .delete()
@@ -109,6 +116,7 @@ const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
       setLike(like + 1);
     }
   };
+
   return (
     <main>
       <div className="px-[92px] mt-[140px]">
@@ -145,7 +153,14 @@ const ConcertDetailPage = ({ params: { postId } }: ConcertDetailPageProps) => {
         <article className="flex gap-[76px]">
           <div className="relative min-w-[450px] w-[450px]">
             {/* Image 태그로 변경 필요 */}
-            {image && <img src={image} alt={title} />}
+            {image && (
+              <Image
+                src={image}
+                alt="공연 포스터 이미지"
+                width={450}
+                height={450}
+              />
+            )}
           </div>
           <div className="w-full flex flex-col text-[#2E2E2E]">
             <div className="mt-[30px] leading-[60px]">
